chore: move log.util into services
